refactor(models): move common objects to separate file
