+++ shared/models/user.ts
@@ -1,21 +1,6 @@
 import { z } from "zod";
 
-export const DateWithParsing = z.preprocess(arg => {
-  if (typeof arg === "string" || arg instanceof Date) return new Date(arg);
-}, z.date());
-export type DateWithParsing = z.infer<typeof DateWithParsing>;
-
-export const TimeFrequencies = z.enum([
-  "daily",
-  "biweekly",
-  "monthly",
-  "bimonthly",
-  "quarterly",
-  "four-monthly",
-  "semi-annually",
-  "annually",
-]);
-export type TimeFrequencies = z.infer<typeof TimeFrequencies>;
+import { Currencies, DateWithParsing, InterestRateTypes, TimeFrequencies } from "./common";
 
 export const User = z.object({
   uuid: z.string().uuid(),
@@ -32,9 +17,9 @@ export const User = z.object({
   profilePicture: z.string().uuid().optional(),
   verificationCode: z.string().optional(),
   verified: z.boolean().default(false),
-  currency: z.enum(["PEN", "USD"]).default("PEN"),
+  currency: Currencies.default("PEN"),
   paymentFrequency: TimeFrequencies.default("monthly"),
-  interestRateType: z.enum(["nominal", "effective"]).default("nominal"),
+  interestRateType: InterestRateTypes.default("nominal"),
   capitalizationType: TimeFrequencies.default("monthly"),
   language: z.string().length(2, "Debe ser un código ISO 639-1 válido").default("es"),
   timezone: z.string().default("America/Lima"),
